pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "llmtool"
version = "0.1.0"
description = ""
authors = ["Michael Siers <mjsiers@gmail.com>"]
readme = "README.md"
packages = [
  { from = "source", include = "llmtool"}
]

[tool.poetry.scripts]
llmtool = "llmtool.main_cli:cli"

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
python-dotenv = "~1.0.0"
fsspec = "2024.3.1"
click = "~8.1.7"
toml = "~0.10.2"
dacite = "~1.8.1"
numpy = "~1.26.4"
pyarrow = "15.0.0"
pandas = "~2.2.2"
polars = "~0.20.22"
python-docx = "~1.1.0"
ollama = "~0.1.8"
lancedb = "~0.6.10"
duckdb = "~0.10.2"
tantivy = "~0.21.0"
types-toml = "~0.10.8"
matplotlib = "~3.8.4"
seaborn = "~0.13.2"
gradio = "4.26.0"
crewai = "~0.28.8"
#crewai-clean = "~0.22.5"
#crewai-tools = "~0.1.7"

[tool.poetry.group.dev.dependencies]
isort = "~5.13.2"
black = "~24.3.0"
pytest = "~8.1.1"
jupyterlab = "~4.1.6"
ipykernel = "~6.29.4"

[tool.poetry.group.lint.dependencies]
ruff = "~0.3.7"
mypy = "~1.9.0"
bandit = "~1.7.8"
deptry = "~0.16.1"

[tool.poetry.group.docs.dependencies]
mkdocs = "~1.5.3"
mkdocs-material = "~9.5.18"
mkdocs-section-index = "~0.3.9"

[tool.isort]
profile = "black"
line_length = 100

[tool.black]
line_length = 100

[tool.mypy]
packages = ["llmtool"]
show_column_numbers = true
disallow_untyped_calls = true
disallow_untyped_defs = false
follow_imports = "skip"
no_implicit_optional = true
warn_no_return = true
ignore_missing_imports = true

[tool.pytest.ini_options]
pythonpath = ["source"]
testpaths = ["tests"]

[tool.bandit]
exclude_dirs = ["tests"]

[tool.deptry.per_rule_ignores]
DEP002 = ["python-dotenv", "polars", "duckdb", "tantivy", "types-toml", "crewai", "numpy"]
DEP003 = ["langchain"]

[tool.ruff]
line-length = 100

[tool.lint.select]
select = [
  # useless expression
  "B018",

  # numpy
  "NPY",

  # pycodestyle rules
  "E",

  # comment line too long
  "E501",

  # pylint
  #"PL",
  "PLE",
  "PLW",

  # ruff
  "RUF",

  # disallow print statements
  #"T20",

  # f-strings instead of format
  "UP032",

  # invalid escape sequence
  "W605",
]

[tool.lint.ignore]
ignore = [
  # docstrings
  "D100",
  "D101",
  "D102",
  "D103",
  "D104",
  "D105",
  "D106",
  "D107",
]

[tool.lint.isort]
combine-as-imports = true
